Conflicts:
	ict_erp/WebContent/WEB-INF/views/member/memberView.jsp
	ict_erp/src/com/ict/erp/controller/MemberServlet.java
	ict_erp/src/com/ict/erp/dao/MenuDAO.java
	ict_erp/src/com/ict/erp/dao/impl/MemberDAOImpl.java
	ict_erp/src/com/ict/erp/dao/impl/MenuDAOImpl.java
	ict_erp/src/com/ict/erp/service/impl/MemberServiceImpl.java
	ict_erp/src/com/ict/erp/vo/MenuInfo.java

+++ ict_erp/src/com/ict/erp/vo/DepartInfo.java
@@ -1,6 +1,6 @@
 package com.ict.erp.vo;
  public class DepartInfo {
-	private int diNum;
+	private Long diNum;
 	private String diCode;
 	private String diName;
 	private String diDesc;
@@ -12,17 +12,17 @@ package com.ict.erp.vo;
 		this.pi = pi;
 	}
 	public DepartInfo() {}
-	public DepartInfo(int diNum, String diCode, String diName, String diDesc) {
+	public DepartInfo(Long diNum, String diCode, String diName, String diDesc) {
 		super();
 		this.diNum = diNum;
 		this.diCode = diCode;
 		this.diName = diName;
 		this.diDesc = diDesc;
 	}
-	public int getDiNum() {
+	public Long getDiNum() {
 		return diNum;
 	}
-	public void setDiNum(int diNum) {
+	public void setDiNum(Long diNum) {
 		this.diNum = diNum;
 	}
 	public String getDiCode() {
